Update naming, fix bugs related to IPC connection, update version number.
