+++ soundtrack/include/LockFreeList.h
@@ -2,6 +2,8 @@
 // Copyright Twitch Interactive, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: GPL-2.0
 
+#pragma once
+
 #include <atomic>
 #include <vector>
